@@ -1,4 +1,8 @@
-/** 
+import createNextIntlPlugin from 'next-intl/plugin'
+
+const withNextIntl = createNextIntlPlugin()
+
+/**
  * Undefined entries are not supported. Push optional patterns to this array only if defined.
  * @type {import('next/dist/shared/lib/image-config').RemotePattern}
  */
@@ -23,6 +27,12 @@ const nextConfig = {
   images: {
     remotePatterns
   },
+  // Required to run in a codespace (see https://github.com/vercel/next.js/issues/58019)
+  experimental: {
+    serverActions: {
+        allowedOrigins: ['localhost:3000'],
+    },
+},
 }
 
-module.exports = nextConfig
+export default withNextIntl(nextConfig)
