@@ -1,29 +1,31 @@
 'use client'
-import { getGroup, getGroupExpenses } from '@/lib/api'
-import { useActiveUser } from '@/lib/hooks'
-import { getTotalActiveUserPaidFor } from '@/lib/totals'
-import { formatCurrency } from '@/lib/utils'
+import { cn, formatCurrency } from '@/lib/utils'
+import { useLocale, useTranslations } from 'next-intl'
 
-type Props = {
-  group: NonNullable<Awaited<ReturnType<typeof getGroup>>>
-  expenses: NonNullable<Awaited<ReturnType<typeof getGroupExpenses>>>
-}
+export function TotalsYourSpendings({
+  totalParticipantSpendings = 0,
+  currency,
+}: {
+  totalParticipantSpendings?: number
+  currency: string
+}) {
+  const locale = useLocale()
+  const t = useTranslations('Stats.Totals')
 
-export function TotalsYourSpendings({ group, expenses }: Props) {
-  const activeUser = useActiveUser(group.id)
-
-  const totalYourSpendings =
-    activeUser === '' || activeUser === 'None'
-      ? 0
-      : getTotalActiveUserPaidFor(activeUser, expenses)
-  const currency = group.currency
+  const balance =
+    totalParticipantSpendings < 0 ? 'yourEarnings' : 'yourSpendings'
 
   return (
     <div>
-      <div className="text-muted-foreground">Total you paid for</div>
+      <div className="text-muted-foreground">{t(balance)}</div>
 
-      <div className="text-lg">
-        {formatCurrency(currency, totalYourSpendings)}
+      <div
+        className={cn(
+          'text-lg',
+          totalParticipantSpendings < 0 ? 'text-green-600' : 'text-red-600',
+        )}
+      >
+        {formatCurrency(currency, Math.abs(totalParticipantSpendings), locale)}
       </div>
     </div>
   )
