@@ -52,12 +52,14 @@ export function useBaseUrl() {
 /**
  * @returns The active user, or `null` until it is fetched from local storage
  */
-export function useActiveUser(groupId: string) {
+export function useActiveUser(groupId?: string) {
   const [activeUser, setActiveUser] = useState<string | null>(null)
 
   useEffect(() => {
-    const activeUser = localStorage.getItem(`${groupId}-activeUser`)
-    if (activeUser) setActiveUser(activeUser)
+    if (groupId) {
+      const activeUser = localStorage.getItem(`${groupId}-activeUser`)
+      if (activeUser) setActiveUser(activeUser)
+    }
   }, [groupId])
 
   return activeUser
